Add a set of string api functions.
Moves the body of the `execute_foo` functions to `impl_foo`, and then
adds a `str_api_foo` function that also calls `impl_foo` but returns a
`char *`.

The ultimate goal is to have `execute_foo` print any output to stdout,
and `str_api_foo` to return it as a string. Code embedding magpie in a
larger program will be expected to call the engine with a single string
as an argument, and parse the response string.

(Right now the code is pretty trivial, and largely redundant, since the
only function returning any output synchronously is `generate`, but we
will gradually provide proper responses for the other functions. For
instance async calls to `simulate` and `autoplay` could return a "pid"
that will let the calling code query their status.)